requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
jinja2==3.1.4
python-multipart==0.0.12
aiofiles==24.1.0
httpx==0.27.2
python-dotenv==1.0.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
Pillow==10.4.0

